refactor(workflow): simplify cohort retry
- isolate durable state and stateless retry operations from the runner
- derive redundant paths and warnings instead of persisting them
- skip seedless base generation and add direct utility coverage
